Tema 02/Langchain Tool/requirements.txt
langchain
langchain-google-community[gmail]
langchain-google-genai
python-dotenv
pydantic
